Add explicit CGFloat casts

+++ components/TextControls/src/Theming/MDCFilledTextField+MaterialTheming.m
@@ -23,7 +23,7 @@ static const CGFloat kTextColorNormalOpacity = (CGFloat)0.87;
 static const CGFloat kNormalLabelColorNormalOpacity = (CGFloat)0.60;
 
 static const CGFloat kPrimaryUnderlineColorNormalOpacity = (CGFloat)0.42;
-static const CGFloat kPrimaryAssistiveLabelColorNormalOpacity = 0.60;
+static const CGFloat kPrimaryAssistiveLabelColorNormalOpacity = (CGFloat)0.60;
 static const CGFloat kPrimaryFloatingLabelColorNormalOpacity = (CGFloat)0.60;
 static const CGFloat kPrimaryFloatingLabelColorEditingOpacity = (CGFloat)0.87;
